+++ ICD.Profound.ConnectPRO/Properties/AssemblyInfo.cs
@@ -4,7 +4,7 @@ using System.Reflection;
 [assembly: AssemblyCompany("ICD Systems")]
 [assembly: AssemblyProduct("ICD.Profound.ConnectPRO")]
 [assembly: AssemblyCopyright("Copyright © ICD Systems 2019")]
-[assembly: AssemblyVersion("7.5.0.0")]
+[assembly: AssemblyVersion("7.6.0.0")]
 
 [assembly: AssemblyInformationalVersion("1.1.0")]
 #if SIMPLSHARP