+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+ - Fixed gauges in theme after ICD.Connect.UI rename
 
 ### Added
  - Showing a confirmation dialog when a remote Zoom participant requests microphone mute